requirements.txt
Flask
neo4j
openai
python-dotenv
